requirements.txt
# aiomysql==0.2.0
# alchemy==20.5
# annotated-types==0.7.0
# anyio==4.8.0
# asgiref==3.8.1
# attrs==25.1.0
# bcrypt==4.2.1
# certifi==2024.12.14
# cffi==1.17.1
# channels==4.2.0
# chardet==3.0.4
# charset-normalizer==3.4.1
# click==8.1.8
# colorama==0.4.6
# crispy-bootstrap5==2024.10
# cryptography==44.0.0
# Django==5.1.5
# django-crispy-forms==2.3
# django-debug-toolbar==5.0.1
# djangorestframework==3.15.2
# dnspython==2.7.0
# docutils==0.21.2
# drf-spectacular==0.28.0
# ecdsa==0.19.0
# email_validator==2.2.0
# fastapi==0.115.6
# fastapi-cli==0.0.7
# filelock==3.0.12
# greenlet==3.1.1
# gunicorn==23.0.0
# h11==0.14.0
# httpcore==1.0.7
# httptools==0.6.4
# httpx==0.28.1
# idna==2.8
# inflection==0.5.1
# itsdangerous==2.2.0
# Jinja2==3.1.5
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# jwt==1.3.1
# lockfile==0.12.2
# mariadb==1.1.11
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# mysql-connector-python==9.1.0
# mysqlclient==2.2.7
# orjson==3.10.15
# packaging==24.2
# passlib==1.7.4
# pillow==11.1.0
# psycopg2==2.9.10
# psycopg2-binary==2.9.10
# pyasn1==0.6.1
# pycparser==2.22
# pydantic==2.10.5
# pydantic-extra-types==2.10.2
# pydantic-settings==2.7.1
# pydantic_core==2.27.2
# Pygments==2.19.1
# PyMySQL==1.1.1
# python-daemon==2.2.4
# python-dotenv==1.0.1
# python-jose==3.3.0
# python-multipart==0.0.20
# pytz==2024.2
# PyYAML==6.0.2
# referencing==0.36.2
# requests==2.32.3
# rich==13.9.4
# rich-toolkit==0.13.2
# rpds-py==0.22.3
# rsa==4.9
# setuptools==75.8.0
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
# sql==2022.4.0
# SQLAlchemy==2.0.37
# sqlparse==0.5.3
# starlette==0.41.3
# typer==0.15.1
# typing_extensions==4.12.2
# tzdata==2024.2
# ujson==5.10.0
# uritemplate==4.1.1
# urllib3==1.25.11
# uvicorn==0.34.0
# vercel==0.2.1
# watchfiles==1.0.4
# websockets==14.2
# wheel==0.45.1
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
python-dotenv
psycopg2-binary